# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gun Del Sol API",
    "description": "High-performance async API for Solana token analysis (Modular)",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/tokens/{token_id}": {
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "Get Token By Id",
        "description": "Get token details with wallets and axiom export",
        "operationId": "get_token_by_id_api_tokens__token_id__get",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tokens"
        ],
        "summary": "Soft Delete Token",
        "description": "Soft delete a token (move to trash)",
        "operationId": "soft_delete_token_api_tokens__token_id__delete",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens/{token_id}/history": {
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "Get Token Analysis History",
        "description": "Get analysis history for a specific token",
        "operationId": "get_token_analysis_history_api_tokens__token_id__history_get",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens/{token_id}/restore": {
      "post": {
        "tags": [
          "Tokens"
        ],
        "summary": "Restore Token",
        "description": "Restore a soft-deleted token",
        "operationId": "restore_token_api_tokens__token_id__restore_post",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens/{token_id}/permanent": {
      "delete": {
        "tags": [
          "Tokens"
        ],
        "summary": "Permanent Delete Token",
        "description": "Permanently delete a token and all associated data",
        "operationId": "permanent_delete_token_api_tokens__token_id__permanent_delete",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/multi-token-wallets": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get Multi Early Buyer Wallets",
        "description": "Get wallets that appear in multiple tokens",
        "operationId": "get_multi_early_buyer_wallets_multi_token_wallets_get",
        "parameters": [
          {
            "name": "min_tokens",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2,
              "title": "Min Tokens"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiTokenWalletsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/wallets/refresh-balances": {
      "post": {
        "tags": [
          "Wallets"
        ],
        "summary": "Refresh Wallet Balances",
        "description": "Refresh wallet balances for multiple wallets (ASYNC)",
        "operationId": "refresh_wallet_balances_wallets_refresh_balances_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBalancesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshBalancesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisHistory": {
        "properties": {
          "token_id": {
            "type": "integer",
            "title": "Token Id"
          },
          "total_runs": {
            "type": "integer",
            "title": "Total Runs"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRun"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "token_id",
          "total_runs",
          "runs"
        ],
        "title": "AnalysisHistory"
      },
      "AnalysisRun": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "analysis_timestamp": {
            "type": "string",
            "title": "Analysis Timestamp"
          },
          "wallets_found": {
            "type": "integer",
            "title": "Wallets Found"
          },
          "credits_used": {
            "type": "integer",
            "title": "Credits Used"
          },
          "wallets": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Wallets"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_timestamp",
          "wallets_found",
          "credits_used",
          "wallets"
        ],
        "title": "AnalysisRun"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Simple message response"
      },
      "MultiTokenWallet": {
        "properties": {
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "token_count": {
            "type": "integer",
            "title": "Token Count"
          },
          "token_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Token Names"
          },
          "token_addresses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Token Addresses"
          },
          "token_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Token Ids"
          },
          "wallet_balance_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wallet Balance Usd"
          }
        },
        "type": "object",
        "required": [
          "wallet_address",
          "token_count",
          "token_names",
          "token_addresses",
          "token_ids",
          "wallet_balance_usd"
        ],
        "title": "MultiTokenWallet"
      },
      "MultiTokenWalletsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "wallets": {
            "items": {
              "$ref": "#/components/schemas/MultiTokenWallet"
            },
            "type": "array",
            "title": "Wallets"
          }
        },
        "type": "object",
        "required": [
          "total",
          "wallets"
        ],
        "title": "MultiTokenWalletsResponse"
      },
      "RefreshBalancesRequest": {
        "properties": {
          "wallet_addresses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Wallet Addresses"
          }
        },
        "type": "object",
        "required": [
          "wallet_addresses"
        ],
        "title": "RefreshBalancesRequest"
      },
      "RefreshBalancesResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RefreshBalancesResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total_wallets": {
            "type": "integer",
            "title": "Total Wallets"
          },
          "successful": {
            "type": "integer",
            "title": "Successful"
          },
          "api_credits_used": {
            "type": "integer",
            "title": "Api Credits Used"
          }
        },
        "type": "object",
        "required": [
          "message",
          "results",
          "total_wallets",
          "successful",
          "api_credits_used"
        ],
        "title": "RefreshBalancesResponse"
      },
      "RefreshBalancesResult": {
        "properties": {
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "balance_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance Usd"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          }
        },
        "type": "object",
        "required": [
          "wallet_address",
          "balance_usd",
          "success"
        ],
        "title": "RefreshBalancesResult"
      },
...
      "TokenDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "token_address": {
            "type": "string",
            "title": "Token Address"
          },
          "token_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Name"
          },
          "token_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Symbol"
          },
          "acronym": {
            "type": "string",
            "title": "Acronym"
          },
          "analysis_timestamp": {
            "type": "string",
            "title": "Analysis Timestamp"
          },
          "first_buy_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Buy Timestamp"
          },
          "wallets_found": {
            "type": "integer",
            "title": "Wallets Found"
          },
          "credits_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Used"
          },
          "last_analysis_credits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Analysis Credits"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "wallets": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Wallets"
          },
          "axiom_json": {
            "items": {},
            "type": "array",
            "title": "Axiom Json"
          }
        },
        "type": "object",
        "required": [
          "id",
          "token_address",
          "token_name",
          "token_symbol",
          "acronym",
          "analysis_timestamp",
          "first_buy_timestamp",
          "wallets_found",
          "wallets",
          "axiom_json"
        ],
        "title": "TokenDetail",
        "description": "Token with wallet details and axiom data"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}